+++ catalog/video/video-mpegts-stream/src/main/java/org/codice/alliance/video/stream/mpegts/metacard/FrameCenterMetacardUpdater.java
@@ -22,11 +22,6 @@ public class FrameCenterMetacardUpdater extends LineStringMetacardUpdater {
     }
 
     @Override
-    public void accept(Visitor visitor) {
-        visitor.visit(this);
-    }
-
-    @Override
     public String toString() {
         return "FrameCenterMetacardUpdater{} " + super.toString();
     }
