+++ catalog/video/video-mpegts-stream/src/main/java/org/codice/alliance/video/stream/mpegts/metacard/LocationCountryCodeMetacardUpdater.java
@@ -1,0 +1,12 @@
+package org.codice.alliance.video.stream.mpegts.metacard;
+
+import org.codice.alliance.libs.klv.AttributeNameConstants;
+
+/**
+ * Set the parent to the union of the child and parent values.
+ */
+public class LocationCountryCodeMetacardUpdater extends UnionMetacardUpdater {
+    public LocationCountryCodeMetacardUpdater() {
+        super(AttributeNameConstants.LOCATION_COUNTY_CODE);
+    }
+}
