+++ catalog/video/video-mpegts-stream/src/main/java/org/codice/alliance/video/stream/mpegts/metacard/MediaEncodingMetacardUpdater.java
@@ -13,50 +13,20 @@
  */
 package org.codice.alliance.video.stream.mpegts.metacard;
 
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
 import org.codice.alliance.libs.klv.AttributeNameConstants;
-
-import ddf.catalog.data.Attribute;
-import ddf.catalog.data.Metacard;
-import ddf.catalog.data.impl.AttributeImpl;
 
 /**
  * Set the media encodings attribute of the parent to be the union of the parent and child.
  */
-public class MediaEncodingMetacardUpdater implements MetacardUpdater {
+public class MediaEncodingMetacardUpdater extends UnionMetacardUpdater {
 
-    private static final String ATTRIBUTE_NAME = AttributeNameConstants.MEDIA_ENCODING;
-
-    @Override
-    public void update(Metacard parent, Metacard child) {
-        Set<Serializable> mediaEncodings = new HashSet<>();
-        addAll(mediaEncodings, parent);
-        addAll(mediaEncodings, child);
-        List<Serializable> list = new ArrayList<>(mediaEncodings);
-        if(!list.isEmpty()) {
-            parent.setAttribute(new AttributeImpl(ATTRIBUTE_NAME, list));
-        }
-    }
-
-    private void addAll(Set<Serializable> serializables, Metacard metacard) {
-        Attribute parentAttribute = metacard.getAttribute(ATTRIBUTE_NAME);
-        if (parentAttribute != null && parentAttribute.getValues() != null) {
-            serializables.addAll(parentAttribute.getValues());
-        }
-    }
-
-    @Override
-    public void accept(Visitor visitor) {
-        visitor.visit(this);
+    public MediaEncodingMetacardUpdater() {
+        super(AttributeNameConstants.MEDIA_ENCODING);
     }
 
     @Override
     public String toString() {
         return "MediaEncodingMetacardUpdater{}";
     }
+
 }
