+++ catalog/video/video-mpegts-stream/src/main/java/org/codice/alliance/video/stream/mpegts/metacard/MetacardUpdater.java
@@ -35,7 +35,7 @@ public interface MetacardUpdater {
 
         void visit(TemporalStartMetacardUpdater temporalStartMetacardUpdater);
 
-        void visit(MediaEncodingMetacardUpdater mediaEncodingMetacardUpdater);
+        void visit(UnionMetacardUpdater unionMetacardUpdater);
     }
 
 }
