+++ catalog/video/video-mpegts-transformer/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -96,6 +96,7 @@
                                 </bean>
                             </argument>
                         </bean>
+                        <bean class="org.codice.alliance.libs.klv.LocationCountryCodeKlvProcessor"/>
                     </list>
                 </argument>
             </bean>
